requirements.txt
streamlit>=1.24.0
opencv-python>=4.7.0
numpy>=1.24.0
face-recognition>=1.3.0
dlib>=19.24.0
pandas>=1.5.0
deepface>=0.0.75
facenet-pytorch>=2.5.2
torch>=2.0.0
mediapipe>=0.10.0
pymongo==4.6.1
python-dotenv==1.0.1 